Post-hackathon cleanup, removed some unnecessary files and organized code a bit.
