+++ Pi/main.py
@@ -1,10 +1,10 @@
 from paho.mqtt import client as mqtt
 import time
 import ssl
-import MQTTDevice
-from CANDataInterpreter import CANInterface
+import MQTT.MQTTDevice
+from can_receive import CANInterface
 
 if __name__ == "__main__":
     can = CANInterface()
-    can.start_receive()
+    can.startReceive()
